Reorganize project with MVC architecture
