+++ src/Controller/ToDoController.java
@@ -1,30 +1,23 @@
-package com.company;
-
+package Controller;
+import Model.TasksList;
+import View.ToDoInterface;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class Menu {
+public class ToDoController {
     private TasksList collection;
     private UserInput taskInput;
+    private ToDoInterface toDoInterface;
 
-    public Menu(){
+    public ToDoController(){
         this.collection = new TasksList();
         this.taskInput = new UserInput();
+        this.toDoInterface = new ToDoInterface();
     }
 
-    public void showMenu() {
-        System.out.println(">> You have X tasks todo and Y tasks are done! ");
-        System.out.println(">> Pick an option:");
-        System.out.println(">> (1) Show Task List (by date or project)");
-        System.out.println(">> (2) Add New Task");
-        System.out.println(">> (3) Edit Task (update, mark as done, remove)");
-        System.out.println(">> (4) Save and Quit ");
-        System.out.println(">>");
-    }
     public void chooseCommand(){
-        System.out.println(">> Welcome to ToDoLy");
         while(true) {
-            showMenu();
+            toDoInterface.showMenu();
             String option = taskInput.getUserInput();
             switch (option) {
                 case "1":
@@ -43,15 +36,13 @@ public class Menu {
                     LocalDate date = LocalDate.parse(deadline, formatter);
                     if(collection.getTaskListSize() == 0) {
                         System.out.println("Your task list is empty. Please create and add a task");
-                    } else if (collection.containsProjectName(date)) {
+                    } else if (collection.containsDateName(date)) {
                         collection.showTasksByDate(date);
                     } else {
                         System.out.println("There are no tasks for this date. Try another one");
                     }
                     break;
                 case "2":
-//                    TasksList collection = new TasksList();
-//                    UserInput task = new UserInput();
                     taskInput.insertDataForTask(collection.getTasksList());
                     break;
                 case "3":
