+++ src/Model/Task.java
@@ -1,4 +1,4 @@
-package com.company;
+package Model;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
