+++ src/Model/TasksList.java
@@ -1,4 +1,5 @@
-package com.company;
+package Model;
+
 import java.time.LocalDate;
 import java.util.ArrayList;
 
@@ -65,7 +66,7 @@ public class TasksList {
        return false;
     }
 
-    public boolean containsProjectName(LocalDate deadline){
+    public boolean containsDateName(LocalDate deadline){
         for(Task task: tasksList){
             if (task.getDeadline().isEqual(deadline)){
                 return true;
