+++ src/View/ToDoInterface.java
@@ -1,0 +1,15 @@
+package View;
+
+public class ToDoInterface {
+
+    public void showMenu() {
+        System.out.println(">> Welcome to ToDoLy");
+        System.out.println(">> You have X tasks todo and Y tasks are done! ");
+        System.out.println(">> Pick an option:");
+        System.out.println(">> (1) Show Task List (by date or project)");
+        System.out.println(">> (2) Add New Task");
+        System.out.println(">> (3) Edit Task (update, mark as done, remove)");
+        System.out.println(">> (4) Save and Quit ");
+        System.out.println(">>");
+    }
+}
